+++ tests/unit/test_run.py
@@ -474,6 +474,29 @@ def test_init_empty(scm_run):
     assert not empty_run.empty
 
 
+def test_empty_meta_cols():
+    empty_run = ScmRun().set_meta("test", 100)
+    assert len(empty_run.meta["test"]) == 0
+    assert empty_run.meta_attributes == [
+        "model",
+        "region",
+        "scenario",
+        "test",
+        "unit",
+        "variable",
+    ]
+
+    empty_run = empty_run.drop_meta("test")
+    assert empty_run.meta_attributes == [
+        "model",
+        "region",
+        "scenario",
+        "unit",
+        "variable",
+    ]
+    assert "test" not in empty_run.meta
+
+
 def test_repr_empty():
     empty_run = ScmRun()
     assert str(empty_run) == empty_run.__repr__()
@@ -2730,10 +2753,20 @@ def test_filter_empty(scm_run, caplog):
     res = empty_run.filter(variable="anything")
     assert res.shape == (0, 3)
 
-    # Filtering for an empty run is a noop (the times aren't filtered)
     res = empty_run.filter(year=range(2000, 2010))
-    assert res.shape == (0, 3)
+    assert res.shape == (0, 1)
     assert id(res) != id(empty_run)
+
+
+def test_filter_empty_times():
+    result = (
+        get_single_ts()
+        .filter(model="nonexistent", log_if_empty=False)
+        .filter(year=[2], log_if_empty=False)
+        .time_points.years()
+    )
+    expected = np.array([2])
+    np.testing.assert_equal(result, expected)
 
 
 @pytest.mark.xfail(